check values vector isnt empty before trying to access it, not safe to assume

+++ src/game.cpp
@@ -140,19 +140,19 @@ void game_select(const dpp::select_click_t &event) {
 		return;
 	}
 	bot.log(dpp::ll_debug, std::to_string(event.command.usr.id) + ": " + custom_id);
-	if (custom_id == "withdraw" && p.in_bank) {
+	if (custom_id == "withdraw" && p.in_bank && !event.values.empty()) {
 		std::vector<std::string> parts = dpp::utility::tokenize(event.values[0], ";");
 		db::query("DELETE FROM game_bank WHERE owner_id = ? AND item_desc = ? AND item_flags = ?", {event.command.usr.id, parts[0], parts[1]});
 		p.possessions.push_back(item{ .name = parts[0], .flags = parts[1] });
 		p.inv_change = true;
 		claimed = true;
-	} else if (custom_id == "deposit" && p.in_bank) {
+	} else if (custom_id == "deposit" && p.in_bank && !event.values.empty()) {
 		std::vector<std::string> parts = dpp::utility::tokenize(event.values[0], ";");
 		db::query("INSERT INTO game_bank (owner_id, item_desc, item_flags ) VALUES(?,?,?)", {event.command.usr.id, parts[0], parts[1]});
 		p.drop_possession(item{ .name = parts[0], .flags = parts[1] });
 		p.inv_change = true;
 		claimed = true;
-	} else if (custom_id == "fight_pvp" && p.in_pvp_picker) {
+	} else if (custom_id == "fight_pvp" && p.in_pvp_picker && !event.values.empty()) {
 		dpp::snowflake user(event.values[0]);
 		challenge_pvp(event, user);
 		claimed = true;
